feat: Refactor dashboard layout and implement new DashboardPage component
